Refactored Enumerator.
